refactor: Unify Panel crossruntime panel
