--- src/Uno.UI/UI/Xaml/Controls/Panel/Panel.skia.cs
@@ -1,90 +1,0 @@
-﻿using Uno.Extensions;
-using Uno.Foundation.Logging;
-using Uno.UI.Controls;
-using System;
-using System.Linq;
-using System.Collections.Generic;
-using System.Text;
-using System.Collections;
-using Uno.UI.DataBinding;
-using Uno.Disposables;
-using Microsoft.UI.Xaml.Data;
-using System.Runtime.CompilerServices;
-using System.Drawing;
-using Uno.UI;
-using Microsoft.UI.Xaml.Media;
-
-using Microsoft.UI.Xaml.Shapes;
-using Uno.UI.Xaml.Controls;
-
-namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class Panel
-	{
-		private readonly BorderLayerRenderer _borderRenderer;
-
-		public Panel()
-		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			Initialize();
-		}
-
-		partial void Initialize();
-
-		partial void UpdateBorder()
-		{
-			// Checking for Window avoids re-creating the layer until it is actually used.
-			if (IsLoaded)
-			{
-				_borderRenderer.Update();
-			}
-		}
-
-		protected virtual void OnChildrenChanged()
-		{
-			UpdateBorder();
-		}
-
-		partial void OnPaddingChangedPartial(Thickness oldValue, Thickness newValue)
-		{
-			UpdateBorder();
-		}
-
-		partial void OnBorderBrushChangedPartial(Brush oldValue, Brush newValue)
-		{
-			UpdateBorder();
-		}
-
-		partial void OnBorderThicknessChangedPartial(Thickness oldValue, Thickness newValue)
-		{
-			UpdateBorder();
-		}
-
-		partial void OnCornerRadiusChangedPartial(CornerRadius oldValue, CornerRadius newValue)
-		{
-			UpdateBorder();
-		}
-
-		/// <summary>        
-		/// Support for the C# collection initializer style.
-		/// Allows items to be added like this 
-		/// new Panel 
-		/// {
-		///    new Border()
-		/// }
-		/// </summary>
-		/// <param name="view"></param>
-		public void Add(UIElement view)
-		{
-			Children.Add(view);
-		}
-
-		protected override void OnBackgroundChanged(DependencyPropertyChangedEventArgs e)
-		{
-			base.OnBackgroundChanged(e);
-			UpdateBorder();
-			UpdateHitTest();
-		}
-	}
-}
